fix 客户端成功接入

+++ src/utils/socket.ts
@@ -30,7 +30,10 @@ export class SocketUpload {
     this.timeGap = timeGap;
   }
   startServer(path: string, onTick: (args: any[]) => void) {
-    fs.unlinkSync(path);
+    const exists = fs.existsSync(path);
+    if (exists) {
+      fs.unlinkSync(path);
+    }
     const server = net.createServer(c => {
       logger.info("客户端成功接入");
       let parser = new SocketParser();
